Properly set vendor roots to the formulas subdirectory
The git repos of formulas contain subdirectories with the actual
salt formulas we want to use. For example, if we have a formula
named 'docker-formula', and the actual formula code resides in
'docker-formula/docker', this is what we want to link to.
Maintaining version back-compatibility we want to link this
subdirectory directly in 'roots' so that it is exposed as
'vendor/root/docker'. This change adopts a strategy of finding
a formulas subdirectories containing the actual formula.

- Look for structure docker-formula/docker
- If not found, link repo directory directly