Sort requirements
Whenever someone does a fab update_formulas (or equivalent) it's
possible (indeed likely) that the order of the requirements in file
might change. This means that any diff is cluttered because context
lines might be included, even for requirements which haven't changed.

This commit keeps the order in alphabet, such that the pertinent changes
are easier to see.